API fixes + remove top level export
